requirements.txt
joblib==1.1.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.1
Pillow==10.0.1
scikit_learn==1.3.1
shap==0.40.0
streamlit==1.7.0
